chore: update next.js version
