+++ crates/pack-tests/tests/snapshot/externals/promise/input/index.ts
@@ -1,4 +1,5 @@
 import $ from "jquery";
 import value, { named } from "async-value";
+import esmValue, { named as esmNamed } from "async-esm-value";
 
-console.log($.ready, value, named);
+console.log($.ready, value.default, value.named, named, esmValue.default, esmValue.__esModule, esmNamed);
