--- crates/pack-tests/tests/snapshot/externals/promise/output/_root-of-the-server___6b612769.js
@@ -1,50 +1,0 @@
-(globalThis["TURBOPACK"] || (globalThis["TURBOPACK"] = [])).push([typeof document === "object" ? document.currentScript : undefined,
-"[externals]/promise-38c1fe80 [external] (promise)", ((__turbopack_context__) => {
-"use strict";
-
-return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
-try {
-var mod = await ($);
-
-__turbopack_context__.v(mod);
-__turbopack_async_result__();
-} catch(e) { __turbopack_async_result__(e); }
-}, true);
-}),
-"[externals]/promise-504fc35f [external] (promise)", ((__turbopack_context__) => {
-"use strict";
-
-return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
-try {
-var mod = await (Promise.resolve({ default: "async-value", named: true }));
-
-__turbopack_context__.v(mod);
-__turbopack_async_result__();
-} catch(e) { __turbopack_async_result__(e); }
-}, true);
-}),
-"[project]/externals/promise/input/index.ts [client] (ecmascript)", ((__turbopack_context__) => {
-"use strict";
-
-return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
-    try {
-        var __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$38c1fe80__$5b$external$5d$__$28$promise$29$__ = __turbopack_context__.i("[externals]/promise-38c1fe80 [external] (promise)");
-        var __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$504fc35f__$5b$external$5d$__$28$promise$29$__ = __turbopack_context__.i("[externals]/promise-504fc35f [external] (promise)");
-        var __turbopack_async_dependencies__ = __turbopack_handle_async_dependencies__([
-            __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$38c1fe80__$5b$external$5d$__$28$promise$29$__,
-            __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$504fc35f__$5b$external$5d$__$28$promise$29$__
-        ]);
-        [__TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$38c1fe80__$5b$external$5d$__$28$promise$29$__, __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$504fc35f__$5b$external$5d$__$28$promise$29$__] = __turbopack_async_dependencies__.then ? (await __turbopack_async_dependencies__)() : __turbopack_async_dependencies__;
-        ;
-        ;
-        console.log(__TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$38c1fe80__$5b$external$5d$__$28$promise$29$__["default"].ready, __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$504fc35f__$5b$external$5d$__$28$promise$29$__["default"], __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$504fc35f__$5b$external$5d$__$28$promise$29$__["named"]);
-        __turbopack_context__.s([]);
-        __turbopack_async_result__();
-    } catch (e) {
-        __turbopack_async_result__(e);
-    }
-}, false);
-}),
-]);
-
-//# sourceMappingURL=_root-of-the-server___6b612769.js.map
