+++ crates/pack-tests/tests/snapshot/externals/promise/output/_root-of-the-server___e03ce6fa.js
@@ -1,0 +1,80 @@
+(globalThis["TURBOPACK"] || (globalThis["TURBOPACK"] = [])).push([typeof document === "object" ? document.currentScript : undefined,
+"[externals]/promise-38c1fe80 [external] (promise)", ((__turbopack_context__) => {
+"use strict";
+
+return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
+try {
+var mod = await ($);
+
+var ns = Object.create(null);
+if (mod && (typeof mod === 'object' || typeof mod === 'function')) {
+  for (var key in mod) ns[key] = mod[key];
+}
+ns.default = mod;
+__turbopack_context__.n(ns);
+__turbopack_async_result__();
+} catch(e) { __turbopack_async_result__(e); }
+}, true);
+}),
+"[externals]/promise-504fc35f [external] (promise)", ((__turbopack_context__) => {
+"use strict";
+
+return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
+try {
+var mod = await (Promise.resolve({ default: "async-value", named: true }));
+
+var ns = Object.create(null);
+if (mod && (typeof mod === 'object' || typeof mod === 'function')) {
+  for (var key in mod) ns[key] = mod[key];
+}
+ns.default = mod;
+__turbopack_context__.n(ns);
+__turbopack_async_result__();
+} catch(e) { __turbopack_async_result__(e); }
+}, true);
+}),
+"[externals]/promise-4821facd [external] (promise)", ((__turbopack_context__) => {
+"use strict";
+
+return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
+try {
+var mod = await (Promise.resolve({ __esModule: true, default: "async-esm-value", named: "esm-named" }));
+
+var ns = Object.create(null);
+if (mod && (typeof mod === 'object' || typeof mod === 'function')) {
+  for (var key in mod) ns[key] = mod[key];
+}
+ns.default = mod;
+__turbopack_context__.n(ns);
+__turbopack_async_result__();
+} catch(e) { __turbopack_async_result__(e); }
+}, true);
+}),
+"[project]/externals/promise/input/index.ts [client] (ecmascript)", ((__turbopack_context__) => {
+"use strict";
+
+return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
+    try {
+        var __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$38c1fe80__$5b$external$5d$__$28$promise$29$__ = __turbopack_context__.i("[externals]/promise-38c1fe80 [external] (promise)");
+        var __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$504fc35f__$5b$external$5d$__$28$promise$29$__ = __turbopack_context__.i("[externals]/promise-504fc35f [external] (promise)");
+        var __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$4821facd__$5b$external$5d$__$28$promise$29$__ = __turbopack_context__.i("[externals]/promise-4821facd [external] (promise)");
+        var __turbopack_async_dependencies__ = __turbopack_handle_async_dependencies__([
+            __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$38c1fe80__$5b$external$5d$__$28$promise$29$__,
+            __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$504fc35f__$5b$external$5d$__$28$promise$29$__,
+            __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$4821facd__$5b$external$5d$__$28$promise$29$__
+        ]);
+        [__TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$38c1fe80__$5b$external$5d$__$28$promise$29$__, __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$504fc35f__$5b$external$5d$__$28$promise$29$__, __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$4821facd__$5b$external$5d$__$28$promise$29$__] = __turbopack_async_dependencies__.then ? (await __turbopack_async_dependencies__)() : __turbopack_async_dependencies__;
+        ;
+        ;
+        ;
+        console.log(__TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$38c1fe80__$5b$external$5d$__$28$promise$29$__["default"].ready, __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$504fc35f__$5b$external$5d$__$28$promise$29$__["default"].default, __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$504fc35f__$5b$external$5d$__$28$promise$29$__["default"].named, __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$504fc35f__$5b$external$5d$__$28$promise$29$__["named"], __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$4821facd__$5b$external$5d$__$28$promise$29$__["default"].default, __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$4821facd__$5b$external$5d$__$28$promise$29$__["default"].__esModule, __TURBOPACK__imported__module__$5b$externals$5d2f$promise$2d$4821facd__$5b$external$5d$__$28$promise$29$__["named"]);
+        __turbopack_context__.s([]);
+        __turbopack_async_result__();
+    } catch (e) {
+        __turbopack_async_result__(e);
+    }
+}, false);
+}),
+]);
+
+//# sourceMappingURL=_root-of-the-server___e03ce6fa.js.map
