+++ crates/pack-tests/tests/snapshot/externals/promise/output/main.js
@@ -1,5 +1,5 @@
 (globalThis["TURBOPACK"] || (globalThis["TURBOPACK"] = [])).push([
     typeof document === "object" ? document.currentScript : undefined,
-    {"otherChunks":["_root-of-the-server___6b612769.js"],"runtimeModuleIds":["[project]/externals/promise/input/index.ts [client] (ecmascript)"]}
+    {"otherChunks":["_root-of-the-server___e03ce6fa.js"],"runtimeModuleIds":["[project]/externals/promise/input/index.ts [client] (ecmascript)"]}
 ]);
 // Dummy runtime
